napari-hub-commons/requirements.txt
gunicorn==20.1.0
pynamodb==5.4.1
Markdown==3.4.1
bs4==0.0.1
cffconvert==2.0.0
pyyaml==6.0
